backend/pyproject.toml
[project]
name = "fastapi-todos"
version = "0.0.1"
description = "Бэкенд веб-приложения Список задач"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.115.8",
    "logfire[fastapi]>=3.5.3",
    "passlib[bcrypt]>=1.7.4",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "pydantic-settings>=2.7.1",
    "pydantic[email]>=2.11.5",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.38",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "alembic>=1.15.1",
    "black>=25.1.0",
    "isort>=6.0.1",
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
